client/requirements.txt
Pillow
picamera